+++ source/bin.ts
@@ -1,0 +1,3 @@
+#!/usr/bin/env node
+import getMAC from './'
+getMAC().then(console.log)